chore: split every page to modules
