+++ package.json
@@ -8,7 +8,7 @@
     "start:local": "http-server -p 8080 -c-1 dist/ch-web",
     "build": "ng build",
     "build:prod": "cross-env NODE_ENV=production ng build --prod",
-    "test": "ng test",
+    "test": "ng test --code-coverage",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
