+++ projects/ch-web/src/app/about/about.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AboutComponent } from './about.component';
+import { RouterModule, Routes } from '@angular/router';
+
+const routes: Routes = [
+  { path: '', component: AboutComponent}
+];
+
+
+@NgModule({
+  declarations: [
+    AboutComponent
+  ],
+  imports: [
+    CommonModule,
+    RouterModule.forChild(routes),
+  ]
+})
+export class AboutModule { }
