+++ projects/ch-web/src/app/app.module.ts
@@ -1,40 +1,25 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MaterialModule } from './material.module';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { DashBoardComponent } from './dash-board/dash-board.component';
-import { AddWatcherComponent } from './add-watcher/add-watcher.component';
-import { AboutComponent } from './about/about.component';
-import { SettingsComponent } from './settings/settings.component';
-import { ReactiveFormsModule } from '@angular/forms';
 import { HttpProgressInterceptor } from './interceptors/http-progress.interceptor';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { environment } from '../environments/environment';
 import { DatePipe } from '@angular/common';
 import { RequestLimiterInterceptor } from './interceptors/request-limiter.interceptor';
-import { WatchDetailsComponent } from './watch-details/watch-details.component';
 import { SharedModule } from './shared/shared.module';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    DashBoardComponent,
-    AddWatcherComponent,
-    AboutComponent,
-    SettingsComponent,        
-    WatchDetailsComponent
+    AppComponent               
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule,
-    HttpClientModule,
-    MaterialModule,
-    ReactiveFormsModule,
+    HttpClientModule,    
     ServiceWorkerModule.register('chweb-service-worker.js', {
       enabled: environment.production,
       // Register the ServiceWorker as soon as the app is stable
