+++ projects/ch-web/src/app/watch/dash-board/dash-board.component.html
@@ -3,7 +3,7 @@
         <div mat-subheader>Active Watches</div>
         <mat-list-item class="my-2" *ngFor="let watch of (activeWatches$ | async); trackBy:trackByWatchId">
             <mat-icon color="accent" mat-list-icon>track_changes</mat-icon>
-            <div mat-line class="cursor-pointer" [routerLink]="['/watch-details', watch.id]">
+            <div mat-line class="cursor-pointer" [routerLink]="['/dashboard/watch-details', watch.id]">
                 <div>
                     <span *ngIf="watch.pin; else displayDistrict" class="font-semibold text-gray-700">Pin:
                         {{watch.pin}}</span>
@@ -12,10 +12,10 @@
                     </ng-template>
                 </div>
             </div>
-            <div mat-line class="text-gray-500 text-base cursor-pointer" [routerLink]="['/watch-details', watch.id]">
+            <div mat-line class="text-gray-500 text-base cursor-pointer" [routerLink]="['/dashboard/watch-details', watch.id]">
                 Last updated: {{watch.lastUpdated | date:'d/M/y, h:mm a'}} 
             </div>
-            <div mat-line class="cursor-pointer" [routerLink]="['/watch-details', watch.id]">
+            <div mat-line class="cursor-pointer" [routerLink]="['/dashboard/watch-details', watch.id]">
                 <div>
                     <span *ngIf="watch?.deltaCenters >= 0; else displayReducedCenters" class="align-text-bottom text-green-700 font-semibold">{{localMath.abs(watch.totalCenters)}} Centers</span>
                     <ng-template #displayReducedCenters>
