+++ projects/ch-web/tsconfig.app.json
@@ -2,8 +2,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../out-tsc/app",
-    "resolveJsonModule": true,
+    "outDir": "../../out-tsc/app",    
     "types": []
   },
   "files": [
